more fixes uffu
